% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{auto_palette}
\alias{auto_palette}
\title{Automatically select a color palette based on signature types}
\usage{
auto_palette(types, default = pal_set2(), verbose = TRUE)
}
\arguments{
\item{types}{A character vector of types e.g. from a sigverse signature 'type' column.}

\item{default}{The default color palette to use if no exact match is found.}

\item{verbose}{verbose (flag)}
}
\value{
A color palette (named vector where names are types and values are colours).
}
\description{
This function takes a vector of signature 'types' and selects an appropriate color palette.
It checks if the types match predefined palettes and
returns the matching palette or a default palette \code{\link[=pal_set2]{pal_set2()}} if no exact match is found.
}
\examples{
\dontrun{# Example usage:
auto_palette(c("T>G", "T>G"))
}

}
