% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_visualise_dimred.R
\name{example_umap}
\alias{example_umap}
\title{Exemplar UMAP dataframe}
\usage{
example_umap()
}
\value{
dataframe with UMAP results
}
\description{
Exemplar UMAP dataframe
}
\examples{
umap <- example_umap()
print(umap)
}
