% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_visualise.R
\name{make_tooltip}
\alias{make_tooltip}
\title{Make a tooltip}
\usage{
make_tooltip(channels, type, mutations1, mutations2, name1, name2, pal)
}
\arguments{
\item{channels}{vector of channel names}

\item{type}{vector of channel types}

\item{mutations1}{mutation counts for catalog 1 (numeric vector)}

\item{mutations2}{mutation counts for catalog 2 (numeric vector)}

\item{name1}{name of sample from which catalog 1 was derived (string)}

\item{name2}{name of sample from which catalog 2 was derived (string)}

\item{pal}{named vector where names are mutation types and values are colours}
}
\value{
a character vector encoding the tooltip to display
}
\description{
Make a tooltip
}
