% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minisig.R
\name{proportion_bar}
\alias{proportion_bar}
\title{Create a Proportion Bar Plot}
\usage{
proportion_bar(
  proportion,
  format = fmt_percent,
  fgcol = "maroon",
  bgcol = "grey90",
  size = NA
)
}
\arguments{
\item{proportion}{Numeric value between 0 and 1 representing the proportion to be plotted.}

\item{format}{Function to format the proportion label. Defaults to \code{\link[=fmt_percent]{fmt_percent()}}.}

\item{fgcol}{Color for the foreground (proportion) part of the bar. Defaults to \code{"maroon"}.}

\item{bgcol}{Color for the background (remaining) part of the bar. Defaults to \code{"grey90"}.}

\item{size}{Numeric value for the text size of the proportion label. Defaults to \code{NA}.}
}
\value{
A ggplot object representing the proportion bar plot.
}
\description{
Creates a horizontal bar plot representing a single proportion, with optional formatting and colors.
}
\examples{
proportion_bar(0.75)
proportion_bar(0.5, fgcol = "blue", bgcol = "lightblue")
}
