% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_visualise.R
\name{sig_make_interactive}
\alias{sig_make_interactive}
\title{Make any sigverse visualisation interactive}
\usage{
sig_make_interactive(gg)
}
\arguments{
\item{gg}{the plot returned from any sigverse visualisation}
}
\value{
a ggiraph interactive visualisation
}
\description{
Make any sigverse visualisation interactive
}
