% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_visualise.R
\name{sig_make_interactive_overlay}
\alias{sig_make_interactive_overlay}
\title{Make sigverse overlay visualisation interactive}
\usage{
sig_make_interactive_overlay(gg)
}
\arguments{
\item{gg}{the plot returned from any sigverse visualisation}
}
\value{
a ggiraph interactive visualisation
}
\description{
Make sigverse overlay visualisation interactive
}
