% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{sig_palette_cn_type}
\alias{sig_palette_cn_type}
\title{Palettes}
\usage{
sig_palette_cn_type()
}
\description{
Colour palettes for common signature types
}
\details{
Palettes are consistent with COSMIC signature database (https://cancer.sanger.ac.uk/signatures/)
}
\examples{
sig_palette_snv_type()
sig_palette_doublet_type()
sig_palette_indel_type()

}
