% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_visualise_bootstraps.R
\name{sig_visualise_bootstraps}
\alias{sig_visualise_bootstraps}
\title{Visualize Bootstrap Contributions for Signatures}
\usage{
sig_visualise_bootstraps(
  bootstraps,
  min_contribution_threshold = 0.05,
  pvalue = 0.05
)
}
\arguments{
\item{bootstraps}{A dataframe in sigverse style describing bootstraps. See \code{\link[sigshared:bootstraps]{sigshared::example_bootstraps()}}.}

\item{min_contribution_threshold}{Numeric value representing the minimum contribution threshold to consider (default is 0.05). See threshold argument of \code{\link[sigstats:sig_compute_experimental_p_value]{sigstats::sig_compute_experimental_p_value()}}.}

\item{pvalue}{The p-value threshold for significance (default is 0.05). P values are computed using \code{\link[sigstats:sig_compute_experimental_p_value]{sigstats::sig_compute_experimental_p_value()}}}
}
\value{
A ggplot object representing the boxplot visualization of bootstrap contributions.
}
\description{
This function takes a dataframe of bootstrap contributions for signatures and visualizes them using boxplots.
It highlights significant contributions based on a p-value threshold and a minimum contribution threshold.
}
\examples{
library(sigshared)
sig_visualise_bootstraps(example_bootstraps())

}
