% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_visualise.R
\name{sig_visualise_compare}
\alias{sig_visualise_compare}
\title{Compare two signature catalogues by overlaying one over the other}
\usage{
sig_visualise_compare(
  catalogue1,
  catalogue2,
  compare = c("count", "fraction"),
  names = c("Sample 1", "Sample 2"),
  channel_order = "auto",
  palette = "auto",
  na.value = "grey",
  subtitle = NULL,
  title = NULL,
  options = vis_options()
)
}
\arguments{
\item{catalogue1, catalogue2}{a sigverse catalogue (tally) representing an observed mutational profile (data.frame)}

\item{compare}{what metric to compare (counts or fractions)}

\item{names}{a 2 element character vector describing the names that should be used to describe catalogue 1 and 2 retrospectively.}

\item{channel_order}{How channels should be ordered on the X axis. By default 'auto' will automatically pick palette if channels are recognised as a standard COSMIC SBS/DBS/INDEL signature, or if not sorts in alphabetical order. Can also be a vector of channels in the order they should appear}

\item{palette}{colours based on the 'type' column. By default 'auto' will automatically pick a palette if the values of the 'type' column matches COSMIC SBS, Doublet or Indel mutations. Otherwise should be a named vector where names = types and values are colours.}

\item{na.value}{colour to use when type = NA}

\item{subtitle}{plot subtitle}

\item{title}{plot title}

\item{options}{other visualisation options. See \code{\link[=vis_options]{vis_options()}} for details}
}
\value{
a ggplot object.
}
\description{
Compare two signature catalogues by overlaying one over the other
}
\examples{
library(sigstats) # For combining signature models
library(TCGAcatalogues) # For pulling example TCGA catalogue data

# Load a catalogue collection (Tally of variant types)
catalogues <- catalogues_load("BRCA", type = "SBS_96")

# Get catalogue of 2 different samples
sample1 = "TCGA-OL-A6VQ-01A-12D-A41F-09"
sample2 = "TCGA-OL-A97C-01A-32D-A41F-09"
catalogue_sample1 <- catalogues[[sample1]]
catalogue_sample2 <- catalogues[[sample2]]


# Compare the two catalogues
ggcompare <- sig_visualise_compare(
  catalogue1 = catalogue_sample1,
  catalogue2 = catalogue_sample2
)

# Make interactive
sig_make_interactive(ggcompare)

}
