% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_visualise_dimred.R
\name{sig_visualise_dimred}
\alias{sig_visualise_dimred}
\title{Visualize Dimensionality Reduction Results}
\usage{
sig_visualise_dimred(
  dimred,
  col_sample = "sample",
  col_dim1 = "dim1",
  col_dim2 = "dim2",
  metadata = NULL,
  col_fill = NULL,
  col_colour = NULL,
  col_shape = NULL,
  title = "Dimensionality Reduction",
  xlab = col_dim1,
  ylab = col_dim2
)
}
\arguments{
\item{dimred}{A data frame containing the dimensionality reduction results with at least columns for the two dimensions
(\code{col_dim1}, \code{col_dim2}) and a sample identifier (\code{col_sample}).}

\item{col_sample}{A string specifying the column name in \code{dimred} that identifies the samples. Default is "sample".}

\item{col_dim1}{A string specifying the column name in \code{dimred} for the first dimension. Default is "dim1".}

\item{col_dim2}{A string specifying the column name in \code{dimred} for the second dimension. Default is "dim2".}

\item{metadata}{Optional. A data frame with metadata corresponding to the samples, with columns for the sample
identifiers (\code{col_sample}) and other optional aesthetic mappings (\code{col_fill}, \code{col_colour}, \code{col_shape}).}

\item{col_fill}{Optional. A string specifying the column name in \code{metadata} for the fill aesthetic. Default is NULL.}

\item{col_colour}{Optional. A string specifying the column name in \code{metadata} for the color aesthetic. Default is NULL.}

\item{col_shape}{Optional. A string specifying the column name in \code{metadata} for the shape aesthetic. Default is NULL.}

\item{title}{A string specifying the title of the plot. Default is "Dimensionality Reduction".}

\item{xlab}{A string specifying the label for the x-axis. Default is the value of \code{col_dim1}.}

\item{ylab}{A string specifying the label for the y-axis. Default is the value of \code{col_dim2}.}
}
\value{
A ggplot object representing the dimensionality reduction plot.
}
\description{
This function creates a scatter plot of the results from a dimensionality reduction technique (e.g., UMAP or PCA).
The plot can include optional metadata features such as color, fill, and shape for
further distinction of sample groups.
}
\examples{
# Example UMAP data
dimred <- example_umap()
metadata <- example_umap_metadata()
sig_visualise_dimred(dimred, metadata = metadata, col_colour = "dataset")
}
