% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_visualise_donut.R
\name{sig_visualise_donut}
\alias{sig_visualise_donut}
\title{Visualize Proportions as a Donut Plot}
\usage{
sig_visualise_donut(
  proportions,
  unexplained_label = "Unexplained",
  inner_radius = 0.6,
  palette = NULL,
  guide_title = NULL
)
}
\arguments{
\item{proportions}{A named numeric vector where the values represent proportions, and names represent labels.}

\item{unexplained_label}{Label for the remaining proportion if proportions sum to less than 1. Defaults to "Unexplained".}

\item{inner_radius}{Radius of the hole in the center of the donut (0 to 1 scale).}

\item{palette}{A named character vector where the names match the \code{proportions} vector and \code{unexplained_label}, and values are colors.}

\item{guide_title}{The legend's title. Set to \code{NULL} for no title.}
}
\value{
A ggplot donut plot.
}
\description{
Visualize Proportions as a Donut Plot
}
\examples{
model = c('SBS2' = 0.6, 'SBS13' = 0.2)
sig_visualise_donut(model)

}
