% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minisig.R
\name{sig_visualise_minified}
\alias{sig_visualise_minified}
\title{Create a Minified Signature Plot with Proportion Bar}
\usage{
sig_visualise_minified(
  signature,
  proportion,
  format = fmt_percent,
  fgcol = "maroon",
  bgcol = "grey90",
  text_size_prop_label = NA,
  heights = c(0.9, 0.1),
  ...
)
}
\arguments{
\item{signature}{A \code{sigverse} style signature data.frame.}

\item{proportion}{Numeric value between 0 and 1 representing the proportion to display.}

\item{format}{Function to format the proportion label. Defaults to \code{\link[=fmt_percent]{fmt_percent()}}.}

\item{fgcol}{Color for the foreground (proportion) part of the bar. Defaults to \code{"maroon"}.}

\item{bgcol}{Color for the background (remaining) part of the bar. Defaults to \code{"grey90"}.}

\item{text_size_prop_label}{Numeric value for the text size of the proportion label in the bar. Defaults to \code{NA}.}

\item{heights}{Numeric vector specifying the relative heights of the signature plot and proportion bar. Defaults to \code{c(0.9, 0.1)}.}

\item{...}{Additional arguments passed to underlying functions.}
}
\value{
A combined ggplot object of the signature plot and proportion bar.
}
\description{
Generates a compact visualization of a signature plot with a proportion bar underneath.
}
\examples{
sig <- sigshared::example_signature()
sig_visualise_minified(sig, proportion = 0.75)

}
