% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_visualise.R
\name{sig_visualise_overlay}
\alias{sig_visualise_overlay}
\title{Visualise a signature overlayed on observed mutational profile}
\usage{
sig_visualise_overlay(signature, catalogue, ...)
}
\arguments{
\item{signature}{a sigverse signature object representing the exposure model (data.frame)}

\item{catalogue}{a sigverse catalogue (tally) representing an observed mutational profile (data.frame)}
}
\value{
a ggplot object.
}
\description{
Visualise a signature overlayed on observed mutational profile
}
\examples{
library(sigstats) # For combining signature models
library(sigstash) # For pulling signatures
library(TCGAcatalogues) # For pulling example TCGA catalogue data

# Load Signature
signatures <- sig_load("COSMIC_v3.3.1_SBS_GRCh38")

# Create a model (combination of signatures)
model <- sig_combine(signatures, model = c('SBS2' = 0.6, 'SBS13' = 0.4))
model_signature <- sig_combine_collapse_to_single_signature(model)

# Load a catalogue (Tally of variant types)
tally <- catalogues_load("BRCA", type = "SBS_96")

# Get tally of a single sample
sample = "TCGA-5L-AAT1-01A-12D-A41F-09"
tally_single_sample <- tally[[sample]]

# Visualise the overlay
gg <- sig_visualise_overlay(
  catalogue = tally_single_sample,
  signature = model_signature
)

# Make interactive
sig_make_interactive(gg)
}
