% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_minisig}
\alias{theme_minisig}
\title{Theme for small signature/catalogue plots}
\usage{
theme_minisig(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[ggplot2:ggtheme]{ggplot2::theme_bw}}
  \describe{
    \item{\code{base_size}}{base font size, given in pts.}
    \item{\code{base_family}}{base font family}
    \item{\code{base_line_size}}{base size for line elements}
    \item{\code{base_rect_size}}{base size for rect elements}
  }}
}
\value{
ggplot2 theme
}
\description{
A minimal theme with no axis text or ticks.
Commonly used for smaller signature/catalogue plots
}
