% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_visualise.R
\name{theme_sigverse}
\alias{theme_sigverse}
\title{Sigverse Theme for Signature Plots}
\usage{
theme_sigverse(
  fontsize_x = NULL,
  fontsize_y = NULL,
  fontsize_title = NULL,
  hjust_title = 0.5,
  ...
)
}
\arguments{
\item{fontsize_x}{fontsize of x axis text (number)}

\item{fontsize_y}{fontsize of y axis text (number)}

\item{fontsize_title}{fontsize of title (number)}

\item{hjust_title}{title horizontal justification (number)}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:ggtheme]{ggplot2::theme_bw}}
  \describe{
    \item{\code{base_size}}{base font size, given in pts.}
    \item{\code{base_family}}{base font family}
    \item{\code{base_line_size}}{base size for line elements}
    \item{\code{base_rect_size}}{base size for rect elements}
  }}
}
\value{
ggplot2 theme
}
\description{
Sigverse Theme for Signature Plots
}
