% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_visualise.R
\name{sig_make_interactive}
\alias{sig_make_interactive}
\title{Make sigverse overlay visualisation interactive}
\usage{
sig_make_interactive(gg, width_svg = 12, height_svg = NULL)
}
\arguments{
\item{gg}{the plot returned from any sigverse visualisation}
}
\value{
a ggiraph interactive visualisation
}
\description{
Make sigverse overlay visualisation interactive
}
