% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_visualise.R
\name{sig_visualise}
\alias{sig_visualise}
\title{Visualise Signatures}
\usage{
sig_visualise(
  signature,
  class = c("signature", "decomposition", "model"),
  title = NULL,
  subtitle = NULL,
  palette = "auto",
  channel_order = "auto",
  na.value = "grey",
  options = vis_options()
)
}
\arguments{
\item{signature}{a sigverse signature object}
}
\value{
ggplot object
}
\description{
Visualises Signatures.
The order channels are displayed are based on the order they appear in the signature data.frame
}
\details{
If the channels match a known channel type, channel display order will instead be based on a predefined order
}
\examples{
library(sigstash)
library(sigstats)
library(TCGAdecomp)

# Load Signature
signatures <- sig_load('COSMIC_v3.3.1_SBS_GRCh38')

# Visualise a single signature
sig_visualise(signatures[["SBS2"]])

# Visualise a decomposition
brca_decompositions <- decomp_load('BRCA')
sig_visualise(brca_decompositions[["TCGA-3C-AALI-01A-11D-A41F-09"]], class = 'decomposition')

# Visualise a model (combination of signatures
model = sig_combine(signatures, model = c('SBS2' = 0.5, 'SBS13' = 0.5))
sig_visualise(model, class = 'model')

# Make Visualisations Interactive
gg = sig_visualise(model, class = 'model')
sig_make_interactive(gg)
}
