% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_visualise.R
\name{theme_sigverse}
\alias{theme_sigverse}
\title{Sigverse Theme for Signature Plots}
\usage{
theme_sigverse(
  fontsize_x = NULL,
  fontsize_y = NULL,
  fontsize_title = NULL,
  hjust_title = 0.5,
  ...
)
}
\value{
ggplot2 theme
}
\description{
Sigverse Theme for Signature Plots
}
