% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_visualise.R
\name{vis_options}
\alias{vis_options}
\title{Set Visualization Options for Signature Plots}
\usage{
vis_options(
  fontsize_x = 6,
  fontsize_y = 9,
  hjust_title = 0.5,
  fontsize_title = 16,
  fontsize_axis_title_y = 14
)
}
\arguments{
\item{fontsize_x}{fontsize of x axis text (number)}

\item{fontsize_y}{fontsize of y axis text (number)}

\item{hjust_title}{title horizontal justification (number)}

\item{fontsize_title}{fontsize of title (number)}

\item{fontsize_axis_title_y}{fontsize of the y axis title (number)}
}
\value{
A list of visualization options.
}
\description{
This function provides a way to set various options for visualizing signature plots.
}
