% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_visualise.R
\name{vis_options}
\alias{vis_options}
\title{Set Visualization Options for Signature Plots}
\usage{
vis_options()
}
\value{
A list of visualization options.
}
\description{
This function provides a way to set various options for visualizing signature plots.
}
